UART version 0.3 DONE
At the current stage the project does perform as expected:
- no TTY-specific bugs;
- total control over both queue RX buffer and temporary input buffer

+++ 01_uart/usartlib/usart_utils.h
@@ -29,10 +29,10 @@ typedef struct {
 	void (*write_char)(char ch);
 } uart_handler_t;
 
-int8_t  uart_open(uint32_t baud, const char* cfg, const char* mode, uint32_t rts, uint32_t cts);
+int32_t  uart_open(uint32_t baud, const char* cfg, const char* mode, uint32_t rts, uint32_t cts);
 int32_t uart_read_keystrokes(char* buf, uint32_t bufsize);
-int8_t uart_getc_nb(void);
-int8_t uart_getc(void);
+int32_t uart_getc_nb(void);
+int32_t uart_getc(void);
 void uart_putc(char ch);
 void uart_puts(const char* buf);
 void USART1_IRQHandler(void);
